Decrease quantity and remove from cart functionality added

+++ Fitness_store/Fitness_store/fitness_app/views.py
@@ -158,7 +158,11 @@ def remove_from_cart(request, product_type, product_id):
 
         cart_item = CartItem.objects.filter(cart=cart, name=product.name).first()
         if cart_item:
-            cart_item.delete()
+            if cart_item.quantity > 1:
+                cart_item.quantity -= 1
+                cart_item.save()
+            else:
+                cart_item.delete()
 
     return redirect('homepage')
